RandIndex: changed size_t to uint64_t

+++ evaluation/RandIndex.h
@@ -22,13 +22,13 @@ private:
 
 	void updateOutputs();
 
-	size_t getNumAgreeingPairs(
+	uint64_t getNumAgreeingPairs(
 			const ImageStack& stack1,
 			const ImageStack& stack2,
-			size_t& numLocations,
-			size_t& numSameComponentPairs1,
-			size_t& numSameComponentPairs2,
-			size_t& numSameComponentPairs12);
+			uint64_t& numLocations,
+			uint64_t& numSameComponentPairs1,
+			uint64_t& numSameComponentPairs2,
+			uint64_t& numSameComponentPairs12);
 
 	// input image stacks
 	pipeline::Input<ImageStack> _reconstruction;
